transient changes while updating code

+++ neon_dev/sgemm/build.sh
@@ -7,8 +7,10 @@ ROOT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")/../.." >/dev/null 2>&1 && pwd)"
 ## Build dependencies
 
 # Ensure Exo is up to date
-(cd "${ROOT_DIR}" && pip uninstall -y exo-lang && python -m build &&
-  pip install dist/*.whl)
+if [ "$1" = "update" ]; then
+  (cd "${ROOT_DIR}" && pip uninstall -y exo-lang && python -m build &&
+     pip install dist/*.whl)
+fi
 
 # set up cmake build
 cmake -S . -B build -DCMAKE_BUILD_TYPE=Release
